% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormDataByScaff.R
\name{getNormDataByScaff}
\alias{getNormDataByScaff}
\title{Obtain a normalized value for each scaffold}
\usage{
getNormDataByScaff(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object that contains normalized
values generated via \code{\link{normalizeCounts}}}
}
\value{
A dataframe with one row per scaffold (or contig) containing a single
representative value for coverage depth in each sample.
}
\description{
The \code{getNormDataByScaff} function generates a single representative
value for normalized coverage depth for each scaffold (or contig) for each
sample.
}
