% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{getNormFactors}
\alias{getNormFactors}
\title{Generate Normalization Factors}
\usage{
getNormFactors(se)
}
\arguments{
\item{se}{a \code{SummarizedExperiment} object created for analysis in ZWYX}
}
\value{
the same \code{SummarizedExperiment} object as input, with additional
column data entries for \code{totalcounts} and \code{windowmedian}
}
\description{
The \code{getNormFactors} function creates normalization factors for each sample in the \code{SummarizedExperiment} object. It creates two types of normalization factors: total counts
of each sample and the median value of counts across windows. These values are
added to the \code{SummarizedExperiment} object that is returned.
}
\details{
This is primarily a utility function used by \code{normalizeCounts} and should
not normally be called directly in routine usage of ZWYX.
}
