% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRatiosByScaff.R
\name{getRatiosByScaff}
\alias{getRatiosByScaff}
\title{Generate Hom:Het ratios per scaffold}
\usage{
getRatiosByScaff(se, XZthr = 0.5, YWthr = -0.5)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object that contains normalized
values generated via \code{\link{normalizeCounts}}}

\item{XZthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being X|Z linked. Scaffolds
with log2(Hom:Het) > \code{XZthr} are considered to be located on the X or Z
chromosome. Default value is 0.5.}

\item{YWthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being Y|W linked. Scaffolds
with log2(Hom:Het) < \code{XZthr} are considered to be located on the Y or W
chromosome. Default value is -0.5.}
}
\value{
A dataframe with four columns corresponding to the sequence (scaffold)
identifier, estimated log2(Hom:Het) ratio, assigned linkage, and sequence
(scaffold) length.
}
\description{
The \code{getRatiosByScaff} function generates a point estimate for each
scaffold (or contig) of the normalized \emph{Homogametic:Heterogametic}
coverage ratio. It also classifies each scaffold as being autosomal, X|Z, or
Y|W based on specified thresholds for the \emph{Hom:Het} ratio.
}
