% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRatiosByScaff.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{Get lengths of each sequence}
\usage{
getSeqLengths(myGRlist)
}
\arguments{
\item{myGRlist}{A GenomicRanges List object}
}
\value{
A vector
}
\description{
A simple utility function that extracts the largest end coordinate for
each sequence in a GenomicRangesList object into a numeric vector.
}
\details{
The working assumption here inside ZWYX is that the GenomicRanges objects
have been prepared such that the largest end coordinate for each unique
sequence is the final base and there indicates the length of that sequence.
}
