% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpUtils.R
\name{makeGRfromCP}
\alias{makeGRfromCP}
\title{Make GenomicRanges from Changepoint}
\usage{
makeGRfromCP(gr, cpt, XZthr = 0.5, YWthr = -0.5)
}
\arguments{
\item{gr}{A \code{GenomicRanges} object that was used as input for running
changepoint analysis.}

\item{cpt}{A \code{changepoint} object.}

\item{XZthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being X|Z linked. Scaffolds
with log2(Hom:Het) > \code{XZthr} are considered to be located on the X or Z
chromosome. Default value is 0.5.}

\item{YWthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being Y|W linked. Scaffolds
with log2(Hom:Het) < \code{XZthr} are considered to be located on the Y or W
chromosome. Default value is -0.5.}
}
\value{
A \code{GenomicRanges} object corresponding to segments detected
by Changepoint analysis.
}
\description{
This function takes as input a genomic ranges object that was used to run
changepoint and returns a genomic ranges object giving the ranges of CP
segments and their associated means.
}
\details{
This is primarily a utility function used by \code{runChangpointOnGR} and should
not normally be called directly in routine usage of ZWYX.
}
