% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSegments.R
\name{mergeSegments}
\alias{mergeSegments}
\title{Merge segments for all chromosomes}
\usage{
mergeSegments(changeList, asDF = FALSE)
}
\arguments{
\item{changeList}{The list output by \code{runChangepointOnSE}}

\item{asDF}{Logical. If TRUE, returns results formatted as a dataframe. The default
value FALSE returns formatted as a GenomicRanges List.}
}
\value{
Either a dataframe or GenomicRanges (default) object containing the
minimum set of ranges needed to indicate autosomale of sex-chromosome linkage
of segments detected by changepoint analysis.
}
\description{
Operating on the list generated by \code{runChangepointOnSE}, this function
merges all adjacent ranges with the linkage type (e.g. Autosomal, X|Z, or
Y|W). It produces the minimum set of coordinate ranges to capture blocks of
linkage inferred by Changepoint analysis.
}
\details{
In many cases, especially for Y|W chromosome, coverage ratios vary substantially
but  still consistently exceed the threshold for being called sex-linked.
This can result in many adjacent segments that are distinguished
by having different means, but are all still considered Y|W linked, for instance.
This function will merged those adjacent segments into a single range, though
the mean log2(Hom:Het) ratio is removed as a result.
}
