% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize Read Counts}
\usage{
normalizeCounts(se, method = "totalreads", scale = NA)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object created for analysis in ZWYX}

\item{method}{The method used for normalization. A character value either
\code{"totalreads"} or \code{"windowmedian"}. If \code{"totalreads"}, each
window's read count in a sample is divided by the total read count across windows
in the sample.
If \code{"windowmedian"}, each window's read count in a sample is divided by
the median of all window counts in the sample.}

\item{scale}{A single numeric value by which normalized window read count values are
multiplied in order rescale values after normalization to be a "pretty" number.
Here "pretty" simply means not an overlay small or large number. Default value
is \code{NA},
which will use \code{1000} for \code{"totalreads"} normalization or \code{1e8}
for \code{"windowmedian"} normalization.}
}
\value{
The function returns a modified version of the same
\code{SummarizedExperiment} (SE) object that was input. The returned SE object
has two additions. First, the column data includes entries for
\code{totalcounts} or \code{windowmedian}, corresponding to which method was
used and generated by the \code{getNormFactors}
function. Second, and more importantly, an additional assay matrix is appended
to the SE object. This appended assay  matrix is named \code{normdata} and
contains the normalized and rescaled read count values, which will be used
for downstream analyses in ZWYX.
}
\description{
The \code{normalizeCounts} function enerates normalized read count values in each window for a
\code{SummarizedExperiment} object. Read counts can be normalized either by total
reads per sample or the median read count across windows. A matrix of normalized, rescaled
read counts is appended as an assay to the input \code{SummarizedExperiment} object.
}
