% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSegments.R
\name{reduceSegments}
\alias{reduceSegments}
\title{Reduce adjacent segments with common linkage}
\usage{
reduceSegments(SegmentGR)
}
\arguments{
\item{SegmentGR}{A GenomicRanges object that is produced from analyzing
sex-linkage with the function \code{runChangepointOnSE}.}
}
\value{
A GenomicRanges object.
}
\description{
This function merges adjacent segments that have the same chromosomal
linkage (i.e. autosomal, X, or Y) into a single unified segment. The main
purpose is to reduce adjacent segments detectd by Changepoint that may have
different means but are still considered to have the same sex-chromosome or
autosomal linkage.
}
