% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChangpoint.R
\name{runChangepointOnSE}
\alias{runChangepointOnSE}
\title{Run Changepoint on All Scaffolds}
\usage{
runChangepointOnSE(se, XZthr = 0.5, YWthr = -0.5)
}
\arguments{
\item{se}{A ZWYX-prepared SummarizedExperiment object containing
normalized read counts.}

\item{XZthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being X|Z linked. Scaffolds
with log2(Hom:Het) > \code{XZthr} are considered to be located on the X or Z
chromosome. Default value is 0.5.}

\item{YWthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being Y|W linked. Scaffolds
with log2(Hom:Het) < \code{XZthr} are considered to be located on the Y or W
chromosome. Default value is -0.5.}
}
\value{
A list with one element for each scaffold. Each element is itself a
list of length four -- the output of \code{runChangpointOnGR}.
For each scaffold the list contains:
\enumerate{
\item The original GR (with ratios in column),
\item GR of segments detected by Changepoint analysis,
\item the Changepoint
object resulting from analysis,
\item a vector of indices indicating which
windows had Inf coverage ratios and thus were removed prior to Changepoint
analysis.
}
}
\description{
This function takes a ZWYX-prepared SummarizedExperiment object containing
normalized read counts and runs changepoint on each scaffold (contig).
}
