% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChangpoint.R
\name{runChangpointOnGR}
\alias{runChangpointOnGR}
\title{Run Changepoint on GenomicRanges}
\usage{
runChangpointOnGR(
  gr,
  gamety,
  method = "PELT",
  penalty = "Hannan-Quinn",
  minseglen = 5,
  XZthr = 0.5,
  YWthr = -0.5,
  ...
)
}
\arguments{
\item{gr}{A GenomicRanges object for a single scaffold, with associated
metadata containg readcounts per window for each sample.}

\item{gamety}{The "gamety" vector from a ZWYX-prepared SummarizedExperiment object.}

\item{method}{Corresponds to argument for \code{cpt.mean} function from Changepoint package}

\item{penalty}{Corresponds to argument for \code{cpt.mean} function from Changepoint package}

\item{minseglen}{Corresponds to argument for \code{cpt.mean} function from Changepoint package}

\item{XZthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being X|Z linked. Scaffolds
with log2(Hom:Het) > \code{XZthr} are considered to be located on the X or Z
chromosome. Default value is 0.5.}

\item{YWthr}{A numeric value corresponding to the log2(Hom:Het) ratio that
determines whether a scaffold is classified as being Y|W linked. Scaffolds
with log2(Hom:Het) < \code{XZthr} are considered to be located on the Y or W
chromosome. Default value is -0.5.}

\item{...}{Other arguments to be provided to \code{cpt.mean} function.}
}
\value{
A list of length 4: 1) The original GR (with ratios in column),
2) GR of segments detected by Changepoint analysis, 3) the Changepoint
object resulting from analysis, 4) a vector of indices indicating which
windows had Inf coverage ratios and thus were removed prior to Changepoint
analysis.
}
\description{
This function runs changepoint on counts from a single scaffold,
provided as GenomicRanges object. It returns a list that contains the input
as well as the result Changepoint results, and any filtered windows.
}
\details{
This is primarily a utility function used by \code{runChangepointOnSE} and should
not normally be called directly in routine usage of ZWYX.
}
