% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormDataByScaff.R
\name{splitGRangesByScaffold}
\alias{splitGRangesByScaffold}
\title{Split \code{GenomicRanges} by Sequence}
\usage{
splitGRangesByScaffold(myGR)
}
\arguments{
\item{myGR}{A \code{GenomicRanges} object.}
}
\value{
A list of \code{GenomicRanges} objects, one per unique sequence
}
\description{
The \code{splitGRangesByScaffold} function takes a \code{GenomicRanges}
object and splits it into a list of \code{GenomicRanges}, one per unique
sequence identified in the input, which typically corresponds to the contig
or scaffold of the genome assembly being analyze with ZWYX.
}
