% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZWYXplot.R
\name{zwyxPlot}
\alias{zwyxPlot}
\title{Plot Chromosomal Segments}
\usage{
zwyxPlot(
  InputGR,
  SegmentGR,
  bpunit = "kbp",
  addLines = TRUE,
  XZLineCol = "dodger blue",
  YWLineCol = "deep pink",
  AutoLineCol = "yellow2",
  seglwd = 1,
  ...
)
}
\arguments{
\item{InputGR}{The GenomicRanges object input to Changepoint analysis,
containing log2(Hom:Het) coverage ratios in metadata}

\item{SegmentGR}{The GenomicRanges object representing the results of Changepoint
analysis, with coordinates of segment boundaries and means as metadata}

\item{bpunit}{A character value, one of "bp", "kbp", or "mbp". This defines
the scale of the plot's X-axis coordinates to be individual base pairs,
kilobases, or megabases, accordingly.}

\item{addLines}{Logical. TRUE will include line segments drawn at the mean
log2(Hom:Het) value for each sequence segment detected by Changepoint}

\item{XZLineCol}{The color for lines indicating Z|X linkage}

\item{YWLineCol}{The color for lines indicating W|Y linkage}

\item{AutoLineCol}{The color for lines indicating autosomal linkage}

\item{seglwd}{Numeric, the line width parameter for lines drawn when
\code{addLines} is TRUE.}

\item{...}{Other standard graphical parameter arguments provided to the \code{plot}
function.}
}
\value{
A base-graphics plot of log2(Hom:Het) ratios in each window, along with
line segments indicating the sequence segments detected by Changepoint.
}
\description{
This function provides a visualiztion of the distinct segments detected by
Changepoint analysis of male:female sequencing coverage ratios.
}
